+++ application/common/ldap/LdapConnectionException.php
@@ -1,6 +1,0 @@
-<?php
-namespace common\ldap;
-
-class LdapConnectionException extends \Exception
-{
-}